backend/requirements.txt
# The Graph API Client Dependencies
httpx>=0.24.0
pydantic>=2.0.0
python-dotenv>=1.0.0

# Testing Dependencies
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.11.0




